frontend/requirements.txt
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
responses==0.24.1
